+++ src/components/Notification/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Notification = (): JSX.Element => <p>Notification</p>;
+
+export default Notification;